feat: #63 남은 시간 계산 hook 생성
