+++ src/components/Home/MeetingListItem.tsx
@@ -7,7 +7,7 @@ import BeerCover from '@/assets/images/ic_beer_cover.svg';
 import DeliveryCover from '@/assets/images/ic_delivery_cover.svg';
 import Clock from '@/assets/images/ic_clock.svg';
 import { flexAlignCenter, flexCenter, flexColumn } from '@/styles/CommonStyle';
-import dayjs from 'dayjs';
+import useRemainingTime from '@/hooks/useRemainingTime';
 interface MeetingListItemProps {
   cover: string;
   isSelected?: boolean;
@@ -123,39 +123,7 @@ const MeetingListItem: React.FC<MeetingListItemProps> = ({
 }) => {
   const coverType =
     cover === 'meal' ? MealCover : cover === 'beer' ? BeerCover : DeliveryCover;
-
-  const [remainingTime, setRemainingTime] = useState<string>('');
-
-  useEffect(() => {
-    const calculateRemainingTime = () => {
-      const now = dayjs();
-      const dueDate = dayjs(time);
-
-      const diff = dueDate.diff(now, 'second');
-      if (diff > 0) {
-        const hours = Math.floor(diff / 3600);
-        const minutes = Math.floor((diff % 3600) / 60);
-        const seconds = diff % 60;
-
-        if (hours > 0) {
-          setRemainingTime(
-            `${hours.toString().padStart(2, '0')}시간 ${minutes.toString().padStart(2, '0')}분`
-          );
-        } else {
-          setRemainingTime(
-            `${minutes.toString().padStart(2, '0')}분 ${seconds.toString().padStart(2, '0')}초`
-          );
-        }
-      } else {
-        setRemainingTime('시간이 만료되었습니다');
-      }
-    };
-
-    calculateRemainingTime();
-    const timer = setInterval(calculateRemainingTime, 1000);
-
-    return () => clearInterval(timer);
-  }, [time]);
+  const remainingTime = useRemainingTime(time);
 
   const formattedDescription =
     description.length > 30 ? `${description.slice(0, 30)}...` : description;
